move input.rs to process_event.rs
