+++ src/context/message_queue.rs
@@ -23,30 +23,27 @@ impl MessageQueue {
         Self::default()
     }
 
-    pub fn push_success(&mut self, msg: String) {
-        let message = Message::new(msg, Style::default().fg(Color::Green));
-        self.push_msg(message);
-    }
-
     pub fn push_info(&mut self, msg: String) {
         let message = Message::new(msg, Style::default().fg(Color::Yellow));
         self.push_msg(message);
     }
-
+    pub fn push_success(&mut self, msg: String) {
+        let message = Message::new(msg, Style::default().fg(Color::Green));
+        self.push_msg(message);
+    }
     pub fn push_error(&mut self, msg: String) {
         let message = Message::new(msg, Style::default().fg(Color::Red));
         self.push_msg(message);
     }
 
-    fn push_msg(&mut self, msg: Message) {
-        self.contents.push_back(msg);
-    }
-
     pub fn pop_front(&mut self) -> Option<Message> {
         self.contents.pop_front()
     }
-
     pub fn current_message(&self) -> Option<&Message> {
         self.contents.front()
+    }
+
+    fn push_msg(&mut self, msg: Message) {
+        self.contents.push_back(msg);
     }
 }
