+++ src/event/mod.rs
@@ -1,3 +1,5 @@
-mod app_event;
+pub mod app_event;
+pub mod process_event;
 
 pub use self::app_event::*;
+pub use self::process_event::*;
