+++ src/util/mod.rs
@@ -2,7 +2,6 @@
 pub mod devicons;
 
 pub mod format;
-pub mod input;
 pub mod keyparse;
 pub mod name_resolution;
 pub mod process;
